Revert: Move RequestImpl back to where it came from
